+++ src/LASR/parser.cpp
@@ -3,6 +3,7 @@
 
 #include "addattribute.h"
 #include "boundaries.h"
+#include "filter.h"
 #include "localmaximum.h"
 #include "noiseivf.h"
 #include "nothing.h"
@@ -205,6 +206,11 @@ bool Pipeline::parse(const SEXP sexpargs, bool build_catalog, bool progress)
           return false; // # nocov
         }
       }
+    }
+    else if (name == "filter")
+    {
+      auto v = std::make_unique<LASRfilter>();
+      pipeline.push_back(std::move(v));
     }
     else if (name == "local_maximum")
     {